requirements.txt
numpy==1.11.2
opencv-python==3.2.0
Pillow==4.2.1
PyQt5==5.9
pyscreenshot==0.4.2
pytesseract==0.1.7
